Extended support for stringified std::function parameters. Templated parameter types are now correctly maintained and referenced. Basic string conversions can be derived from stringified templates. Added handling for unnamed arguments in function headers. Fixed incomplete preprocessor definitions in autogenerated files. Added legible argument names to all function types where available.

+++ targets/lua/templates/lambda.c
@@ -21,7 +21,7 @@ if (${out_value}Func != 0)
 
         auto S = LuaEngine::getInstance()->getLuaStack();
         #if $ret_type.name != "void"
-        return S->executeFunction(${arg_count});
+        return (int)S->executeFunction(${arg_count});
         #else
         S->executeFunction(${arg_count});
         #end if
